+++ internal/mwdd/files/embed/mediawiki/MwddSettings.php
@@ -137,10 +137,11 @@ if(!isset($dockerDbType)){
 if(!isset($dockerDbType)){
 	$message = "<html><head></head><body>" . PHP_EOL;
 	$message .= "<h1>mwcli: Unable to find database</h1>" . PHP_EOL;
-	$message .= '<p>Could not find a running database for the database name <pre>' . $dockerDb . '</pre></p>' . PHP_EOL;
+	$message .= '<p>Could not find a running database for the database name: <pre>' . $dockerDb . '</pre></p>' . PHP_EOL;
 	$message .= '<p>Please ensure that the site is installed and the database service chosen is running.</p>' . PHP_EOL;
-	$message .= '<p>You can check running services with <pre>mw docker docker-compose ps</pre></p>' . PHP_EOL;
-	$message .= '<p>You can install a new site with <pre>mw docker mediawiki install</pre></p>' . PHP_EOL;
+	$message .= '<p>You can check running services with: <pre>mw docker docker-compose ps</pre></p>' . PHP_EOL;
+	$message .= '<p>You can check installed sites with: <pre>mw docker mediawiki sites</pre></p>' . PHP_EOL;
+	$message .= '<p>You can install a new site with: <pre>mw docker mediawiki install</pre></p>' . PHP_EOL;
 	$message .= '</body></html>' . PHP_EOL;
 	die($message);
 }